Using Distance Sensor and Motor
In this project, we learned how to connect the Control Hub to a computer and program both a motor and a distance sensor. To do this, we created a new branch in Android Studio, placed it in the correct folder, and wrote the necessary code. We also renamed the motors and sensors on the devices to match our program, then tested the code, debugged a few errors, and ran it successfully.

Overall, this project introduced us to some Java basics for working with motors and sensors. By using sample code as a reference, we were able to make the motor respond to the left joystick on the gamepad and have the distance sensor detect objects.

Changes:
- speed of motor turn based on force of joystick push